layouts (css flexbox) + Better setScreen + block component

+++ Core/src/main/kotlin/dev/lynith/core/ui/styles/impl/Border.kt
@@ -3,7 +3,7 @@ package dev.lynith.core.ui.styles.impl
 import dev.lynith.core.ui.styles.Style
 
 class Border @JvmOverloads constructor(
-    var thickness: Int = 0,
+    var thickness: Float = 0f,
     var color: Color = Color(0)
 ) : Style<Border>() {
